Added save and addyear functionality
We're done! (if it all works).
Added a function to the filehandler that overwrites the currently open file with the information currently stored by the program.
Added a new form where you can input all the data needed to create a new year.
Added a method to the location class to allow it to accept a new year and catalogue it correctly.